Move SimpleRangeValue from ./src/interpreter to ./src so that the typedoc documentation is generated for it
