+++ src/ArrayValue.ts
@@ -6,7 +6,7 @@
 import {ArraySize} from './ArraySize'
 import {CellError} from './Cell'
 import {EmptyValue, InternalScalarValue, InterpreterValue} from './interpreter/InterpreterValue'
-import {SimpleRangeValue} from './interpreter/SimpleRangeValue'
+import {SimpleRangeValue} from './SimpleRangeValue'
 
 export interface IArray {
   size: ArraySize,
