+++ src/Lookup/ColumnBinarySearch.ts
@@ -7,7 +7,7 @@ import {SimpleCellAddress} from '../Cell'
 import {CellValueChange} from '../ContentChanges'
 import {DependencyGraph} from '../DependencyGraph'
 import {RawNoErrorScalarValue, RawScalarValue} from '../interpreter/InterpreterValue'
-import {SimpleRangeValue} from '../interpreter/SimpleRangeValue'
+import {SimpleRangeValue} from '../SimpleRangeValue'
 import {ColumnsSpan} from '../Span'
 import {AdvancedFind} from './AdvancedFind'
 import {ColumnSearchStrategy, SearchOptions} from './SearchStrategy'
