+++ src/Lookup/SearchStrategy.ts
@@ -8,7 +8,7 @@ import {Config} from '../Config'
 import {CellValueChange} from '../ContentChanges'
 import {DependencyGraph} from '../DependencyGraph'
 import {RawInterpreterValue, RawNoErrorScalarValue, RawScalarValue} from '../interpreter/InterpreterValue'
-import {SimpleRangeValue} from '../interpreter/SimpleRangeValue'
+import {SimpleRangeValue} from '../SimpleRangeValue'
 import {ColumnsSpan} from '../Span'
 import {Statistics} from '../statistics'
 import {ColumnBinarySearch} from './ColumnBinarySearch'
