+++ src/SimpleRangeValue.ts
@@ -3,12 +3,13 @@
  * Copyright (c) 2022 Handsoncode. All rights reserved.
  */
 
-import {AbsoluteCellRange} from '../AbsoluteCellRange'
-import {ArraySize} from '../ArraySize'
-import {CellError, ErrorType, simpleCellAddress, SimpleCellAddress} from '../Cell'
-import {DependencyGraph} from '../DependencyGraph'
-import {ErrorMessage} from '../error-message'
-import {InternalScalarValue, isExtendedNumber} from './InterpreterValue'
+import {AbsoluteCellRange} from './AbsoluteCellRange'
+import {ArraySize} from './ArraySize'
+import {CellError, ErrorType, simpleCellAddress, SimpleCellAddress} from './Cell'
+import {DependencyGraph} from './DependencyGraph'
+import {ErrorMessage} from './error-message'
+import {InternalScalarValue, isExtendedNumber} from './interpreter/InterpreterValue'
+
 
 export class SimpleRangeValue {
   public readonly size: ArraySize
