+++ src/interpreter/plugin/DateTimePlugin.ts
@@ -30,7 +30,7 @@ import {
   RawNoErrorScalarValue,
   RawScalarValue,
 } from '../InterpreterValue'
-import {SimpleRangeValue} from '../SimpleRangeValue'
+import {SimpleRangeValue} from '../../SimpleRangeValue'
 import {FunctionArgumentType, FunctionPlugin, FunctionPluginTypecheck} from './FunctionPlugin'
 
 /**
